+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Reporting about omitted elements which are flagged as obligatory.
     * Reporting about missmatched type.
     * Aliasing via `HookupAlias` property is supported.
+* Introduced custom abbreviations to enhance code completions capabilities. Head over to [README#custom-abbreviations](README.md#custom-abbreviations) to learn how to use them.
 
 ## [0.10.1] - 2019-06-20